Move participant address selector to participantSelectors file
