+++ spec/javascripts/selectors/participantSelectorsSpec.js
@@ -5,6 +5,7 @@ import {
   selectParticipants,
   selectRoles,
   selectAllRoles,
+  selectFormattedAddresses,
 } from 'selectors/participantSelectors'
 import * as matchers from 'jasmine-immutable-matchers'
 
@@ -183,6 +184,73 @@ describe('participantSelectors', () => {
       ])
       const state = Map().set('participants', participants)
       expect(selectAllRoles(state)).toEqualImmutable(List())
+    })
+  })
+
+  describe('selectFormattedAddresses', () => {
+    it('returns info for the person with the passed id', () => {
+      const people = [
+        {id: '1', addresses: [{type: 'Residence'}]},
+        {id: '2', addresses: [{type: 'Cell'}]},
+      ]
+      const state = fromJS({participants: people})
+      expect(selectFormattedAddresses(state, '1').first().get('type'))
+        .toEqual('Residence')
+    })
+
+    it('returns an empty array if no addresses exists for the person', () => {
+      const people = [{id: '1'}]
+      const state = fromJS({participants: people})
+      expect(selectFormattedAddresses(state, '1')).toEqualImmutable(List())
+    })
+
+    it('returns the street for an address', () => {
+      const people = [{id: '1', addresses: [{street: '1234 Nowhere Lane'}]}]
+      const state = fromJS({participants: people})
+      expect(selectFormattedAddresses(state, '1').first().get('street'))
+        .toEqual('1234 Nowhere Lane')
+    })
+
+    it('returns the city for an address', () => {
+      const people = [{id: '1', addresses: [{city: 'Somewhereville'}]}]
+      const state = fromJS({participants: people})
+      expect(selectFormattedAddresses(state, '1').first().get('city'))
+        .toEqual('Somewhereville')
+    })
+
+    it('returns the formatted state for an address', () => {
+      const people = [{id: '1', addresses: [{state: 'CA'}]}]
+      const state = fromJS({participants: people})
+      expect(selectFormattedAddresses(state, '1').first().get('state'))
+        .toEqual('California')
+    })
+
+    it('returns an empty string for an invalid state', () => {
+      const people = [{id: '1', addresses: [{state: ''}]}]
+      const state = fromJS({participants: people})
+      expect(selectFormattedAddresses(state, '1').first().get('state'))
+        .toEqual('')
+    })
+
+    it('returns the zip for an address', () => {
+      const people = [{id: '1', addresses: [{zip: '12345'}]}]
+      const state = fromJS({participants: people})
+      expect(selectFormattedAddresses(state, '1').first().get('zip'))
+        .toEqual('12345')
+    })
+
+    it('returns the zip errors for an address', () => {
+      const people = [{id: '1', addresses: [{zip: '1234'}]}]
+      const state = fromJS({participants: people})
+      expect(selectFormattedAddresses(state, '1').first().get('zipError'))
+        .toEqual(['zip code should be 5 digits'])
+    })
+
+    it('returns the type for an address', () => {
+      const people = [{id: '1', addresses: [{type: 'Residence'}]}]
+      const state = fromJS({participants: people, addressTypes: [{value: 'Residence'}]})
+      expect(selectFormattedAddresses(state, '1').first().get('type'))
+        .toEqual('Residence')
     })
   })
 })
