+++ spec/javascripts/selectors/screening/personShowSelectorsSpec.js
@@ -3,7 +3,6 @@ import {
   getFormattedPersonInformationSelector,
   getFormattedPersonWithErrorsSelector,
   getPersonFormattedPhoneNumbersSelector,
-  getAllPersonFormattedAddressesSelector,
   getReadOnlyPersonFormattedAddressesSelector,
   getNamesRequiredSelector,
   getPersonAlertErrorMessageSelector,
@@ -250,73 +249,6 @@ describe('personShowSelectors', () => {
       ]
       const state = fromJS({participants: people})
       expect(getReadOnlyPersonFormattedAddressesSelector(state, '1').size).toEqual(1)
-    })
-  })
-
-  describe('getAllPersonFormattedAddressesSelector', () => {
-    it('returns info for the person with the passed id', () => {
-      const people = [
-        {id: '1', addresses: [{type: 'Residence'}]},
-        {id: '2', addresses: [{type: 'Cell'}]},
-      ]
-      const state = fromJS({participants: people, systemCodes})
-      expect(getAllPersonFormattedAddressesSelector(state, '1').first().get('type'))
-        .toEqual('Residence')
-    })
-
-    it('returns an empty array if no addresses exists for the person', () => {
-      const people = [{id: '1'}]
-      const state = fromJS({participants: people})
-      expect(getAllPersonFormattedAddressesSelector(state, '1')).toEqualImmutable(List())
-    })
-
-    it('returns the street for an address', () => {
-      const people = [{id: '1', addresses: [{street: '1234 Nowhere Lane'}]}]
-      const state = fromJS({participants: people})
-      expect(getAllPersonFormattedAddressesSelector(state, '1').first().get('street'))
-        .toEqual('1234 Nowhere Lane')
-    })
-
-    it('returns the city for an address', () => {
-      const people = [{id: '1', addresses: [{city: 'Somewhereville'}]}]
-      const state = fromJS({participants: people})
-      expect(getAllPersonFormattedAddressesSelector(state, '1').first().get('city'))
-        .toEqual('Somewhereville')
-    })
-
-    it('returns the formatted state for an address', () => {
-      const people = [{id: '1', addresses: [{state: 'CA'}]}]
-      const state = fromJS({participants: people})
-      expect(getAllPersonFormattedAddressesSelector(state, '1').first().get('state'))
-        .toEqual('California')
-    })
-
-    it('returns an empty string for an invalid state', () => {
-      const people = [{id: '1', addresses: [{state: ''}]}]
-      const state = fromJS({participants: people})
-      expect(getAllPersonFormattedAddressesSelector(state, '1').first().get('state'))
-        .toEqual('')
-    })
-
-    it('returns the zip for an address', () => {
-      const people = [{id: '1', addresses: [{zip: '12345'}]}]
-      const state = fromJS({participants: people})
-      expect(getAllPersonFormattedAddressesSelector(state, '1').first().get('zip'))
-        .toEqual('12345')
-    })
-
-    it('returns the zip errors for an address', () => {
-      const people = [{id: '1', addresses: [{zip: '1234'}]}]
-      const state = fromJS({participants: people})
-      expect(getAllPersonFormattedAddressesSelector(state, '1').first().get('zipError'))
-        .toEqual(['zip code should be 5 digits'])
-    })
-
-    it('returns the type for an address', () => {
-      const people = [{id: '1', addresses: [{type: 'Residence'}]}]
-      const state = fromJS({participants: people, addressTypes: [{value: 'Residence'}]})
-      expect(getAllPersonFormattedAddressesSelector(state, '1').first().get('type'))
-        .toEqual('Residence')
     })
   })
